+++ src/nonAxisAligned.h
@@ -25,10 +25,10 @@ public:
 	 */
 	nonAxisAligned(Vector2D position, float direction, RectSize size) :
 		DrawableObject(position, direction, size),
-		_pi{atan(1) * 4}
+		_pi{(const float) (atan(1) * 4)}
 	{
-		_cornerDistance = sqrt(pow(_size.Height/2,2) + pow(_size.Width/2,2));
-		_cornerAngle = atan(_size.Height/_size.Width) * 180 / _pi;
+		_cornerDistance = (float) sqrt(pow(_size.Height / 2, 2) + pow(_size.Width / 2, 2));
+		_cornerAngle = (float) (atan(_size.Height / _size.Width) * 180 / _pi);
 	}
 
 	//! Virtual function - Returns the front left corner of the object.
@@ -37,10 +37,10 @@ public:
 	 */
 	virtual Vector2D getFrontLeft(){
 		// Calculate the location of the front left corner of the object
-		float xLocation = DrawableObject::getPosition().x -
-				_cornerDistance * cos((DrawableObject::getDirection() - _cornerAngle) * _pi / 180);
-		float yLocation = DrawableObject::getPosition().y -
-				_cornerDistance * sin((DrawableObject::getDirection() - _cornerAngle) * _pi / 180);
+		float xLocation = (float) (DrawableObject::getPosition().x -
+                        _cornerDistance * cos((DrawableObject::getDirection() - _cornerAngle) * _pi / 180));
+		float yLocation = (float) (DrawableObject::getPosition().y -
+                        _cornerDistance * sin((DrawableObject::getDirection() - _cornerAngle) * _pi / 180));
 
 		return Vector2D{xLocation, yLocation};
 	}
@@ -51,10 +51,10 @@ public:
 	 */
 	virtual Vector2D getFrontRight(){
 		// Calculate the location of the front right corner of the tank
-		float xLocation = DrawableObject::getPosition().x -
-				_cornerDistance * cos((DrawableObject::getDirection() + _cornerAngle) * _pi / 180);
-		float yLocation = DrawableObject::getPosition().y -
-				_cornerDistance * sin((DrawableObject::getDirection() + _cornerAngle) * _pi / 180);
+		float xLocation = (float) (DrawableObject::getPosition().x -
+                        _cornerDistance * cos((DrawableObject::getDirection() + _cornerAngle) * _pi / 180));
+		float yLocation = (float) (DrawableObject::getPosition().y -
+                        _cornerDistance * sin((DrawableObject::getDirection() + _cornerAngle) * _pi / 180));
 
 		return Vector2D{xLocation, yLocation};
 	}
@@ -65,10 +65,10 @@ public:
 	 */
 	virtual Vector2D getBackLeft(){
 		// Calculate the location of the back left corner of the tank
-		float xLocation = DrawableObject::getPosition().x -
-				_cornerDistance * cos((180 + DrawableObject::getDirection() + _cornerAngle) * _pi / 180);
-		float yLocation = DrawableObject::getPosition().y -
-				_cornerDistance * sin((180 + DrawableObject::getDirection() + _cornerAngle) * _pi / 180);
+		float xLocation = (float) (DrawableObject::getPosition().x -
+                        _cornerDistance * cos((180 + DrawableObject::getDirection() + _cornerAngle) * _pi / 180));
+		float yLocation = (float) (DrawableObject::getPosition().y -
+                        _cornerDistance * sin((180 + DrawableObject::getDirection() + _cornerAngle) * _pi / 180));
 
 		return Vector2D{xLocation, yLocation};
 	}
@@ -79,10 +79,10 @@ public:
 	 */
 	virtual Vector2D getBackRight(){
 		// Calculate the location of the back right corner of the tank
-		float xLocation = DrawableObject::getPosition().x -
-				_cornerDistance * cos((180 + DrawableObject::getDirection() - _cornerAngle) * _pi / 180);
-		float yLocation = DrawableObject::getPosition().y -
-				_cornerDistance * sin((180 + DrawableObject::getDirection() - _cornerAngle) * _pi / 180);
+		float xLocation = (float) (DrawableObject::getPosition().x -
+                        _cornerDistance * cos((180 + DrawableObject::getDirection() - _cornerAngle) * _pi / 180));
+		float yLocation = (float) (DrawableObject::getPosition().y -
+                        _cornerDistance * sin((180 + DrawableObject::getDirection() - _cornerAngle) * _pi / 180));
 
 		return Vector2D{xLocation, yLocation};
 	}
